Moved the last rooms function over to the RoomsReadModel and got rid of the rooms altogether.
